refactor(cache): move FSCache to a dedicated file
